[DSv4][P1-5] Align FP32 gamma inputs and refresh golden hashes
Keep the RMSNorm gain in FP32 as confirmed with the maintainer. Previously
the BF16 storage path discarded gain bits before the operator's FP32 math.

Implementation:
- Change NormParams validation and seeded fixtures to FP32 gamma.
- Read gamma through float pointers in CUDA forward and input-gradient
  kernels; update Triton dtype validation without changing reduction order.
- Preserve BF16 activations, upstream gradients, normalized/residual outputs,
  and FP32 explicit dx/dgamma. Autograd returns BF16 x.grad and FP32 gamma.grad.
- Update gtest/benchmark input generation and record benchmark gamma dtype.
- Adapt direct callers to the oracle backward signature without x; keep
  provider/backend interfaces with explicit x through the reference adapter.
- Add tests using gamma=1+2**-10 to detect premature BF16 rounding in both
  forward and backward, plus dtype validation and gradient-precision checks.

Golden migration:
- Regenerate the CPU-oracle manifest for the approved FP32 input contract.
- Update 59 hashes: normalized outputs and gamma-dependent upstream
  gradients in five block fixtures, plus y/grad.dx in rms_edges.
- Leave Sinkhorn-edge, residual, and dgamma hashes unchanged and retain
  strict manifest equality; no candidate output is used as a golden value.
- The migration validation recorded in the original refresh commit rounded
  gamma through BF16 in a diagnostic process and reproduced the entire old
  manifest. No diagnostic rounding was added to production code or tests.

H100 PCIe validation (PyTorch 2.8.0+cu128, Triton 3.4.0, CUDA 12.8):
- uv run --no-sync python -m pytest tests/test_rmsnorm_residual.py
  tests/test_rmsnorm_residual_triton.py -q -rs: 28 passed in 7.89s.
- uv run --no-sync python -m pytest
  tests/test_rmsnorm_residual_gamma_precision.py -q: 4 passed in 2.63s.
- uv run --no-sync python -m pytest tests/test_p1_oracle.py
  tests/test_p1_provider.py tests/test_operator_inputs.py
  tests/test_kernel_registry.py -q: 77 passed in 2.16s, including the anchor.
- check_operator.py for cuda and triton at T=16/D=4096 with --check-grad:
  both passed, zero reported errors, gamma gradient FP32.
- check_p1.py for both providers: all boundaries byte-equal in all five cases.
- Benchmark at T=32768/131072/262144, D=4096, 50 warmup/1000 iterations:
  both candidates byte-equal to the oracle before timing; CUDA combined
  speedup 1.27-1.31x versus torch-native on these measured workloads.

Squashes the FP32 implementation and reviewed golden refresh into one
atomic contract migration. The source tree is unchanged from 6b24500.

+++ rl_engine/kernels/ops/triton/rmsnorm_residual_triton.py
@@ -53,7 +53,7 @@ def _check_forward(x, gamma, eps) -> tuple[int, int]:
         raise ValueError(f"T must be in [1, {_MAX_T}]")
     if d not in _VALID_D:
         raise ValueError(f"D must be one of {_VALID_D}")
-    _check_tensor(gamma, "gamma", dtype=torch.bfloat16, shape=(d,), device=x.device)
+    _check_tensor(gamma, "gamma", dtype=torch.float32, shape=(d,), device=x.device)
     if float(eps) != _EPS:
         raise ValueError(f"eps must be exactly {_EPS}")
     return t, d
